requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-json-widget==2.0.1
django-mass-edit==3.5.0
django-mathfilters==1.0.0
djangoql==0.18.1
djangorestframework==3.15.1
ply==3.11
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.1
